test: move fixture helper functions to dedicated file
